+++ src/util/parse-package-name.js
